Tests for acapy/credo with anoncreds format
